+++ source/js/options.ts
@@ -1,4 +1,7 @@
 import optionsStorage from './options-storage';
+import {popupOnly} from './context/popupOnly';
+
+popupOnly('options');
 
 void optionsStorage.syncForm('#options-form');